# EES-1730 Add initial performance optimizations for charts/tables