--- a/src/GovUk.Education.ExploreEducationStatistics.Data.Services/ViewModels/ObservationViewModel.cs
+++ b/src/GovUk.Education.ExploreEducationStatistics.Data.Services/ViewModels/ObservationViewModel.cs
@@ -1,7 +1,5 @@
 using System.Collections.Generic;
-using GovUk.Education.ExploreEducationStatistics.Common.Model;
 using GovUk.Education.ExploreEducationStatistics.Common.Model.Data;
-using GovUk.Education.ExploreEducationStatistics.Data.Model;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
@@ -12,13 +10,13 @@ namespace GovUk.Education.ExploreEducationStatistics.Data.Services.ViewModels
         public IEnumerable<string> Filters { get; set; }
 
         // TODO could this be [JsonConverter(typeof(EnumToEnumValueJsonConverter<GeographicLevel>))]?
-        [JsonConverter(typeof(StringEnumConverter))]
+        [JsonConverter(typeof(StringEnumConverter), true)]
         public GeographicLevel GeographicLevel { get; set; }
 
         public LocationViewModel Location { get; set; }
 
         public Dictionary<string, string> Measures { get; set; }
-        
+
         public string TimePeriod { get; set; }
     }
 }