--- a/src/explore-education-statistics-admin/src/pages/release/datablocks/components/ReleaseDataBlocksPageTabs.tsx
+++ b/src/explore-education-statistics-admin/src/pages/release/datablocks/components/ReleaseDataBlocksPageTabs.tsx
@@ -293,6 +293,7 @@ const ReleaseDataBlocksPageTabs = ({
                 title="Table"
                 key="table"
                 id="manageDataBlocks-table"
+                lazy
               >
                 <TableTabSection
                   table={response.table}