--- a/src/explore-education-statistics-common/src/modules/table-tool/components/WizardStep.tsx
+++ b/src/explore-education-statistics-common/src/modules/table-tool/components/WizardStep.tsx
@@ -55,9 +55,7 @@ const WizardStep = ({ children, id, ...restProps }: WizardStepProps) => {
         </span>
 
         {typeof children === 'function'
-          ? children({
-              ...injectedWizardProps,
-            })
+          ? children(injectedWizardProps)
           : children}
       </div>
     </li>