… src/explore-education-statistics-common/src/modules/charts/util/__tests__/createDataSetCategories.test.ts
… src/explore-education-statistics-common/src/modules/charts/util/__tests__/groupResultMeasuresByDataSet.test.ts
… src/explore-education-statistics-common/src/modules/table-tool/components/__tests__/DownloadCsvButton.test.tsx
… src/explore-education-statistics-common/src/modules/table-tool/components/__tests__/TimePeriodDataTable.test.tsx
… src/explore-education-statistics-common/src/modules/table-tool/components/__tests__/__data__/TimePeriodDataTable.data.tsx
… src/explore-education-statistics-common/src/modules/table-tool/components/utils/__tests__/groupResultsMeasuresByCombination.test.ts