This PR adds the following performance optimizations:

- Adds pre-grouping of result measures by their respective set of filters. This drastically speeds up lookup of results during `createDataSetCategories` (up to ~10x faster)
- Adds a similar pre-grouping of result measures optimization to `TimePeriodDataTable` as well (~4-5x faster)
- Aggregates multiple operations required for calculating `MapBlock` geometries into a single operation (~0.25x faster)
- Removes client-side camelCasing of `geographicLevel` for both charts and tables. This only shaves off a fairly neglible amount of time (~1%), but it mainly makes sense to push any camelCasing onto the server for consistency.

## Relevant changes

- Fixes `ReleaseDataBlockSourceWizardFinalStep` rendering its table twice unnecessarily.
- Changes 'Table' tab in `ReleaseDataBlocksPageTabs` to be rendered lazily i.e. will only be rendered when the user visits the tab. This should help mitigate potentially slow initial renders.

## Benchmark notes

I benchmarked by creating two data blocks from absence data. They were created with map charts using the following filters:

- Block 1: all local authorities, 5 time periods, 3 indicators  and 1 filter
- Block 2: all local authorities, 1 time period, 3 indicators and 2 school types

### Baseline

- Block 1
  - Map: 2566.8ms
  - Table: 428.6ms
- Block 2
  - Map: 174ms
  - Table: 174ms

### After optimizations

- Block 1
  - Map: 199.9 - 203.3ms
  - Table: 91.5ms
- Block 2
  - Map: 101.2 - 114.8ms
  - Table: 43ms

## Further work

There are most likely some more minor performance tweaks that we can do to the main bottlenecked areas (`createDataSetCategories` and `TimePeriodDataTable`), however, these may not have a very good benefit/effort ratio. These may only yield 10-20% speed ups for quite a lot of refactoring and general decrease in readability.

The crux of the issue is that for large tables/charts, there are just a lot of iterations that have to be performed as there is a lot of data to crunch! Additionally, a fairly large chunk of execution time is out of our control as it's up to Leaflet to actually render the maps.

A better investment of time would be to try and remove/reduce UI freezes so that users can't really perceive any slowness. Ideally we want to offload any processing to another thread before the table/chart renders. Web workers will most likely be needed for this (see [EES-1809](https://dfedigital.atlassian.net/browse/EES-1809)).